Move Test To New Folder, Rewrite Test To Account For ApiParser
